pyproject.toml
[project]
name = "digital-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.29.0",
    "cohere>=5.20.0",
    "fastapi[standard]>=0.115.6",
    "google-genai>=0.2.0",
    "httpx>=0.26.0",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.1.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "sentence-transformers>=2.2.2",
    "uvicorn[standard]>=0.27.0",
]
